requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
colorama==0.4.4
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.45.0
django-autoslug==1.9.8
idna==3.2
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
oauthlib==3.1.1
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.19
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
